move the topological sort to prequery generator
