+++ webapi/src/main/scala/org/knora/webapi/messages/util/search/gravsearch/prequery/AbstractPrequeryGenerator.scala
@@ -28,6 +28,8 @@ import org.knora.webapi.messages.util.search.gravsearch.types._
 import org.knora.webapi.messages.v2.responder.valuemessages.DateValueContentV2
 import org.knora.webapi.messages.{OntologyConstants, SmartIri, StringFormatter}
 import org.knora.webapi.util.ApacheLuceneSupport.LuceneQueryString
+import scalax.collection.Graph
+import scalax.collection.GraphEdge.DiHyperEdge
 
 import scala.collection.mutable
 
@@ -2090,5 +2092,111 @@ abstract class AbstractPrequeryGenerator(constructClause: ConstructClause,
       case _ => true
     }
     optimisedPatterns
+  }
+
+  private def createAndSortGraph(statementPatterns: Seq[StatementPattern]): Seq[QueryPattern] = {
+    def makeGraphWithoutCycles(graphComponents: Seq[(String, String)]): Graph[String, DiHyperEdge] = {
+      val graph = graphComponents.foldLeft(Graph.empty[String, DiHyperEdge]) { (graph, edgeDef) =>
+        val edge = DiHyperEdge(edgeDef._1, edgeDef._2)
+        graph + edge // add nodes and edges to graph
+      }
+      val acyclicGraph = if (graph.isCyclic) {
+        // get the cycle
+        val cycle: graph.Cycle = graph.findCycle.get
+        // the cyclic node is the one that cycle starts and ends with
+        val cyclicNode: graph.NodeT = cycle.endNode
+        val cyclicEdge: graph.EdgeT = cyclicNode.edges.last
+        val originNodeOfCyclicEdge: String = cyclicEdge._1.value
+        val TargetNodeOfCyclicEdge: String = cyclicEdge._2.value
+        val graphComponenetsWithOutCycle =
+          graphComponents.filterNot(edgeDef => edgeDef.equals((originNodeOfCyclicEdge, TargetNodeOfCyclicEdge)))
+
+        makeGraphWithoutCycles(graphComponenetsWithOutCycle)
+      } else {
+        graph
+      }
+      acyclicGraph
+    }
+    def createGraph: Graph[String, DiHyperEdge] = {
+      val graphComponents: Seq[(String, String)] = statementPatterns.map { statementPattern =>
+        // transform every statementPattern to pair of nodes that will consist an edge.
+        val node1 = statementPattern.subj.toSparql
+        val node2 = statementPattern.obj.toSparql
+        (node1, node2)
+      }
+      makeGraphWithoutCycles(graphComponents)
+    }
+    def sortStatementPatterns(createdGraph: Graph[String, DiHyperEdge],
+                              statementPatterns: Seq[StatementPattern]): Seq[QueryPattern] = {
+      // Try topological sorting of graph
+      val topologicalOrderSeq: Seq[createdGraph.TopologicalOrder[createdGraph.NodeT]] =
+        createdGraph.topologicalSort match {
+          // Is there still a cycle in the graph?
+          case Left(cycleNode) =>
+            // Don't try sorting, return statements as they are given.
+            Seq.empty[createdGraph.TopologicalOrder[createdGraph.NodeT]]
+          case Right(topOrder) =>
+            // No. return the topological order
+            Seq(topOrder)
+        }
+
+      // Is there a topological order found?
+      val sortedPatterns = if (topologicalOrderSeq.nonEmpty) {
+        // Topological sort algorithm return only one perturbation; i.e. one topologicalOrder.
+        val topologicalOrder: Iterable[createdGraph.NodeT] = topologicalOrderSeq.head
+        // Start from the end of the ordered list (the nodes with lowest degree);
+        // for each node, find statements which have the node as object and bring them to top.
+        val sortedStatements: Seq[QueryPattern] =
+          topologicalOrder.foldRight(Vector.empty[QueryPattern]) { (node, sortedStatements) =>
+            val statementsOfNode: Set[QueryPattern] = statementPatterns
+              .filter(p => p.obj.toSparql.equals(node.value))
+              .toSet[QueryPattern]
+            sortedStatements ++ statementsOfNode.toVector
+          }
+        sortedStatements
+      } else {
+        statementPatterns
+      }
+      sortedPatterns
+    }
+
+    val createdGraph = createGraph
+    sortStatementPatterns(createdGraph, statementPatterns)
+
+  }
+
+  protected def reorderPatternsByDependency(patterns: Seq[QueryPattern]): Seq[QueryPattern] = {
+
+    val statementPatterns: Seq[StatementPattern] = patterns.collect { case pattern: StatementPattern => pattern }
+    val sortedStatementPatterns: Seq[QueryPattern] = createAndSortGraph(statementPatterns)
+    val otherPatterns: Seq[QueryPattern] = patterns.filterNot(p => p.isInstanceOf[StatementPattern])
+    val sortedOtherPatterns: Seq[QueryPattern] = otherPatterns.map {
+      // sort statements inside each UnionPattern block
+      case unionPattern: UnionPattern => {
+        val sortedUnionBlocks: Seq[Seq[QueryPattern]] =
+          unionPattern.blocks.map(block => reorderPatternsByDependency(block).toSeq)
+        UnionPattern(blocks = sortedUnionBlocks)
+      }
+      // sort statements inside OptionalPattern
+      case optionalPattern: OptionalPattern => {
+        val sortedOptionalPatterns: Seq[QueryPattern] = reorderPatternsByDependency(optionalPattern.patterns)
+        OptionalPattern(patterns = sortedOptionalPatterns.toSeq)
+      }
+      // sort statements inside MinusPattern
+      case minusPattern: MinusPattern => {
+        val sortedMinusPatterns: Seq[QueryPattern] = reorderPatternsByDependency(minusPattern.patterns)
+        MinusPattern(patterns = sortedMinusPatterns.toSeq)
+      }
+      // sort statements inside FilterNotExistsPattern
+      case filterNotExistsPattern: FilterNotExistsPattern => {
+        val sortedFilterNotExistsPatterns: Seq[QueryPattern] =
+          reorderPatternsByDependency(filterNotExistsPattern.patterns)
+        FilterNotExistsPattern(patterns = sortedFilterNotExistsPatterns)
+      }
+      // return any other query pattern as it is
+      case pattern: QueryPattern => pattern
+    }
+    val sorted = sortedStatementPatterns ++ sortedOtherPatterns
+    sorted
   }
 }
